refactor(mapgen): group domain authoring ownership
