+++ docs/system/libs/mapgen/explanation/DOMAIN-MODELING.md
@@ -142,8 +142,8 @@ Preferred dependency direction:
 
 ## Ground truth anchors
 
-- Domain contract authoring: `packages/mapgen-core/src/authoring/domain.ts`
-- Domain router authoring: `packages/mapgen-core/src/authoring/domain-router.ts`
+- Domain contract authoring: `packages/mapgen-core/src/authoring/domain/contract.ts`
+- Domain router authoring: `packages/mapgen-core/src/authoring/domain/router.ts`
 - Op authoring helpers: `packages/mapgen-core/src/authoring/op/contract.ts`
 - Example aggregate contract/router: `mods/mod-swooper-maps/src/domain/foundation/contract.ts`, `mods/mod-swooper-maps/src/domain/foundation/router.ts`
 - Example direct module: `mods/mod-swooper-maps/src/domain/foundation/modules/tectonics/`
