+++ packages/mapgen-core/src/authoring/domain/index.ts
@@ -1,0 +1,16 @@
+export type {
+  DomainAggregateContract,
+  DomainAggregateContractAny,
+  DomainContract,
+  DomainContractAny,
+  DomainSubdomainContract,
+  DomainSubdomainContractAny,
+} from "./contract.js";
+export { defineDomain, defineDomainSubdomain } from "./contract.js";
+export type {
+  DomainModule,
+  DomainOpImplementationsForContracts,
+  DomainRouter,
+  DomainSubdomainRouter,
+} from "./router.js";
+export { createDomain, createDomainRouter, createDomainSubdomainRouter } from "./router.js";
